refactor: remove dead code for stops, modify data transmission

+++ userinput-stm32/lib/transmitter/transmitter.h
@@ -5,6 +5,6 @@
 #include <stdint.h>
 
 void transmitter_init(void);
-int transmitter_sendselections(uint16_t stop_idx, uint32_t lines_selected);
+int transmitter_sendselections(uint16_t stop_idx, uint16_t lines_selected);
 
 #endif // __TRANSMITTER__H
